nettle: refresh the patches
* fixes:
WARNING: nettle-2.7.1-r0 do_patch:
Some of the context lines in patches were ignored. This can lead to incorrectly applied patches.
The context lines in the patches can be updated with devtool:

    devtool modify <recipe>
    devtool finish --force-patch-refresh <recipe> <layer_path>

Then the updated patches and the source tree (in devtool's workspace)
should be reviewed to make sure the patches apply in the correct place
and don't introduce duplicate lines (which can, and does happen
when some of the context is ignored). Further information:
http://lists.openembedded.org/pipermail/openembedded-core/2018-March/148675.html
https://bugzilla.yoctoproject.org/show_bug.cgi?id=10450
Details:
Applying patch Add-target-to-only-build-tests-not-run-them.patch
patching file Makefile.in
Hunk #1 succeeded at 53 (offset -2 lines).
patching file testsuite/Makefile.in
Hunk #1 succeeded at 105 with fuzz 2 (offset -11 lines).

Now at patch Add-target-to-only-build-tests-not-run-them.patch

Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>