bash: Provide /bin/{sh, bash} when usrmerge is used
Most shell scripts have '#!/bin/{sh,bash}' on the first line of the
script, which triggers RPM to automatically add a runtime dependency
on that path for any package that contains shell scripts.  However,
when the usrmerge feature is enabled, the path will actually be
/usr/bin/{sh,bash}.

So, to satisfy the runtime dependencies, add '/bin/{sh,bash}' to what
the bash package provides.

Signed-off-by: Peter Kjellerstedt <peter.kjellerstedt@axis.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>