conf/distro: Add disable-gplv3.inc
We're finding meta-gplv2 needs configuration to work as intended. Rather
than teaching this to things like the project autobuilder, collect
the configuration inside an include file in the layer itself which
everyone can either use directly or refer to.

Initial population is from the autobuilder config currently used for
testing meta-gplv2.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>