layer.conf: Delete as the layer is now officially obsolete and unmaintained
Remove the layer.conf as this layer is no longer being maintained and
we need to draw people's attention to that.

The software in this layer has not seen security fixes in a long time.
As such, nobody should be using it. There are other ways of removing GPLv3
components from modern OE/YP builds and we'd like to focus people's
attention onto those. If there are issues that current solutions can't
meet, we need to find new ways to address them, this layer isn't it.

We've talked about this layers maintenance for a long time on the mailing
lists and in weekly tech calls so this shouldn't surprise anyone.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>